Remove long wait by adding grasp closure time (#200)
The [video in the Pick and Place tutorial](http://docs.ros.org/kinetic/api/moveit_tutorials/html/doc/pick_place/pick_place_tutorial.html) is 45 seconds long. For 28 of those 45 seconds, the robot is standing still. This is because a default wait time of 7 seconds is added after every opening and closing of the gripper. This change fixes that time to 0.5 s and makes the movement look smooth.

I also propose notifying the user about this behavior in [this PR](https://github.com/ros-planning/moveit/pull/1004).